--- a/murmur128.go
+++ b/murmur128.go
@@ -31,11 +31,15 @@ type digest128 struct {
 	h2 uint64 // Unfinalized running hash part 2.
 }
 
-func New128() Hash128 {
-	d := new(digest128)
+func SeedNew128(seed1, seed2 uint64) Hash128 {
+	d := &digest128{h1: seed1, h2: seed2}
 	d.bmixer = d
 	d.Reset()
 	return d
+}
+
+func New128() Hash128 {
+	return SeedNew128(0, 0)
 }
 
 func (d *digest128) Size() int { return 16 }
@@ -170,19 +174,19 @@ func fmix64(k uint64) uint64 {
 	return k
 }
 
-/*
-func rotl64(x uint64, r byte) uint64 {
-	return (x << r) | (x >> (64 - r))
-}
-*/
-
 // Sum128 returns the MurmurHash3 sum of data. It is equivalent to the
 // following sequence (without the extra burden and the extra allocation):
 //     hasher := New128()
 //     hasher.Write(data)
 //     return hasher.Sum128()
 func Sum128(data []byte) (h1 uint64, h2 uint64) {
-	d := &digest128{h1: 0, h2: 0}
+	return SeedSum128(0, 0, data)
+}
+
+// SeedSum128 returns the MurmurHash3 sum of data with digests initialized to
+// seed1 and seed2.
+func SeedSum128(seed1, seed2 uint64, data []byte) (h1 uint64, h2 uint64) {
+	d := &digest128{h1: seed1, h2: seed2}
 	d.tail = d.bmix(data)
 	d.clen = len(data)
 	return d.Sum128()