--- a/murmur128_decl.go
+++ b/murmur128_decl.go
@@ -1,0 +1,7 @@
+// +build go1.5,amd64
+
+package murmur3
+
+func Sum128(data []byte) (h1 uint64, h2 uint64)
+
+func SeedSum128(seed1, seed2 uint64, data []byte) (h1 uint64, h2 uint64)