--- a/murmur32_decl.go
+++ b/murmur32_decl.go
@@ -3,3 +3,5 @@
 package murmur3
 
 func Sum32(data []byte) (h1 uint32)
+
+func SeedSum32(seed uint32, data []byte) (h1 uint32)