@twmb 2016-05-11
Tip again down to negligible improvements; I'll evaluate this when 1.7 is released.